+++ shoop/admin/modules/contacts/views/edit.py
@@ -8,8 +8,6 @@
 
 from __future__ import unicode_literals
 
-import ast
-
 from django import forms
 from django.contrib.auth import get_user_model
 from django.db.models.loading import get_model
@@ -21,9 +19,8 @@ from django.utils.translation import ugettext_lazy as _
 from shoop.admin.form_part import (
     FormPart, FormPartsViewMixin, SaveFormPartsMixin, TemplatedFormDef
 )
-from shoop.admin.forms.widgets import (
-    PersonContactChoiceWidget, Select2Multiple
-)
+from shoop.admin.forms.fields import Select2MultipleField
+from shoop.admin.forms.widgets import PersonContactChoiceWidget
 from shoop.admin.toolbar import get_default_edit_toolbar
 from shoop.admin.utils.urls import get_model_url
 from shoop.admin.utils.views import CreateOrUpdateView
@@ -86,17 +83,12 @@ class ContactBaseForm(BaseModelForm):
 
     def formfield_callback(self, f, **kwargs):
         if str(f) == "shoop.CompanyContact.members":
-            initial = PersonContact.objects.none()
+            formfield = Select2MultipleField(
+                model=PersonContact, required=False, help_text=f.help_text)
             if hasattr(self.instance, "members"):
-                initial = self.instance.members.all()
-            k = {
-                "initial": initial,
-                "required": False,
-                "help_text": f.help_text,
-                "widget": Select2Multiple(model=PersonContact)
-            }
-            formfield = forms.CharField(**k)
-            formfield.widget.choices = [(object.pk, force_text(object)) for object in initial]
+                formfield.widget.choices = [
+                    (object.pk, force_text(object)) for object in self.instance.members.all()
+                ]
             return formfield
         return f.formfield(**kwargs)
 
@@ -149,12 +141,6 @@ class ContactBaseForm(BaseModelForm):
     def _clean_fields(self):
         super(ContactBaseForm, self)._clean_fields()
         self.set_model_from_cleaned_data()
-
-    def clean_members(self):
-        members = self.cleaned_data.get("members", [])
-        if members:
-            members = ast.literal_eval(members)
-        return members
 
     def save(self, commit=True):
         obj = super(ContactBaseForm, self).save(commit)